requirements.txt
datasets
ipykernel
jupyter
notebook
openai
python-dotenv